lab4/requirements.txt
fastapi 
uvicorn 
python-dotenv 
